Hoist logging config object, so we can insert enclave_level
