+++ doc/host_config_schema/cchost_config.json
@@ -535,6 +535,12 @@
           "default": "Info",
           "description": "Logging level for the untrusted host."
         },
+        "enclave_level": {
+          "type": "string",
+          "enum": ["Trace", "Debug", "Info", "Fail", "Fatal"],
+          "default": "Info",
+          "description": "Logging level for the enclave. The minimum logging level may be restricted at build-time, in which case this value may be ignored. For instance, SGX release enclaves will remove information-leaking Debug logging at compile-time."
+        },
         "format": {
           "type": "string",
           "enum": ["Text", "Json"],
@@ -673,12 +679,6 @@
       "type": "integer",
       "description": "Cap at which node-to-node message channels will be closed, and a new channel will be created. Can be used to limit use of single cryptographic key",
       "minimum": 0
-    },
-    "enclave_logging_level": {
-      "type": "string",
-      "enum": ["Trace", "Debug", "Info", "Fail", "Fatal"],
-      "default": "Info",
-      "description": "Logging level for the enclave. The minimum logging level may be restricted at build-time, in which case this value may be ignored. For instance, SGX release enclaves will remove information-leaking Debug logging at compile-time."
     }
   },
   "required": ["enclave", "network", "command"],
