+++ include/ccf/node/startup_config.h
@@ -15,6 +15,17 @@
 #include <string>
 #include <vector>
 
+namespace ccf::config
+{
+  enum class LogFormat
+  {
+    TEXT,
+    JSON
+  };
+  DECLARE_JSON_ENUM(
+    LogFormat, {{LogFormat::TEXT, "Text"}, {LogFormat::JSON, "Json"}});
+}
+
 struct CCFConfig
 {
   size_t worker_threads = 0;
@@ -25,8 +36,15 @@ struct CCFConfig
   consensus::Configuration consensus = {};
   ccf::NodeInfoNetwork network = {};
 
-  // TODO: Should we move currently host-only logging object here?
-  logger::Level enclave_logging_level = logger::Level::INFO;
+  struct Logging
+  {
+    logger::Level host_level = logger::Level::INFO;
+    logger::Level enclave_level = logger::Level::INFO;
+    ccf::config::LogFormat format = ccf::config::LogFormat::TEXT;
+
+    bool operator==(const Logging&) const = default;
+  };
+  Logging logging = {};
 
   struct NodeCertificateInfo
   {
