+++ tests/config.jinja
@@ -68,9 +68,9 @@
   "logging":
   {
     "host_level": "{{ host_log_level }}",
+    "enclave_level": "{{ enclave_log_level }}",
     "format": "Text"
   },
-  "enclave_logging_level": "{{ enclave_log_level }}",
   "consensus":
   {
     "message_timeout": "{{ message_timeout }}",
